release-20.1:Backport!3725

See merge request noc/noc!3728

+++ sa/profiles/HP/ProCurve/get_interfaces.py
@@ -72,7 +72,7 @@ class Script(BaseScript):
                 )
 
         i = 0
-        for s in range(len(lines) / step):
+        for s in range(int(len(lines) / step)):
             for str in lines[i : i + step]:
                 leaf = str.split(".")[0]
                 val = str.split("=")[1].lstrip()
